Fix integration test

+++ services/py-api/src/server/schemas/jwt_schemas/jwt_user_data_schema.py
@@ -18,4 +18,4 @@ class JwtUserData(BaseTypedDict):
     team_name: str
     team_id: str
     is_invite: bool
-    exp: int
+    exp: float
